% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_Excel_num.R
\name{Excel_all_numeric}
\alias{Excel_all_numeric}
\title{Show data in spreadsheet
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
Excel_all_numeric(data)
}
\arguments{
\item{data}{a \code{data.frame}, a \code{tibble} or any other object as defined in \code{?openxlsx::writeData()}.}
}
\value{
This function is called for its side-effect of opening a spreadsheet.
}
\description{
Quickly inspect your data in a spreadsheet with auto-filters an all columns formatted as numbers. Currently, this is a very tight definition of a helper function. More flexibility is planned in the future.
}
\examples{
# Excel_all_numeric(airquality) # works fine with a data.frame
# Excel_all_numeric(WorldPhones) # but ignores row names (e.g. from arrays)
}
\references{
A wrapper around the brilliant R package \code{openxlsx}:
\url{https://ycphs.github.io/openxlsx/index.html}
}
\author{
Harald Kaba (2021)
}
