% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excel_write.R
\name{Excel_write_numeric}
\alias{Excel_write_numeric}
\title{Write data to spreadsheet
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
Excel_write_numeric(data)
}
\arguments{
\item{data}{a \code{data.frame}, a \code{tibble} or any other object as defined in \code{?openxlsx::writeData()}.}
}
\value{
This function is primarily called for its side-effect of writing a file. Returns \code{TRUE} if successful, \code{FALSE} otherwise.
}
\description{
Quickly write your data into a spreadsheet with auto-filters and all columns formatted as numbers. Currently, this is a very tight definition of a helper function. More flexibility is planned in the future.
}
\examples{
\dontrun{
Excel_write_numeric(airquality) # works fine with a data.frame
Excel_write_numeric(WorldPhones) # but ignores row names (e.g. from arrays)}
}
\references{
A wrapper around the brilliant R package \code{openxlsx}:
\url{https://ycphs.github.io/openxlsx/index.html}
}
\author{
Harald Kaba (2021)
}
