% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{hist}
\alias{hist}
\title{Histogram and statistics}
\usage{
hist(data, plot_x = NULL, plot_y = NULL, alpha_1 = 0.05, alpha_2 = 0.95)
}
\arguments{
\item{data}{dataframe}

\item{plot_x}{x (defaults to second column in dataframe)}

\item{plot_y}{y}

\item{alpha_1}{lower/left probability}

\item{alpha_2}{higher/right probability}
}
\value{
Histogram
}
\description{
Histogram and statistics
}
\examples{
hist(data = mtcars, plot_x = "wt")
}
