% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piano.R
\name{piano}
\alias{piano}
\title{Piano Plot}
\usage{
piano(data, plot_x = NULL)
}
\arguments{
\item{data}{dataframe}

\item{plot_x}{x (defaults to second column in dataframe)}
}
\value{
A piano plot of the data.
}
\description{
The piano plot, introduced in this function on Feb 15, 2022, is based on the violin plot and, likewise, supposed to improve information density of the plot.
}
\examples{
piano(data.frame(a = 1, x = runif(100))) # a is an arbitrary first column
}
