% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what.R
\name{what}
\alias{what}
\title{What is this Data Frame?}
\usage{
what(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
A skim_df object, see package \code{skimr}.
}
\description{
Print the output of some exploratory helper functions to console.
}
\examples{
HK::what(mtcars)
}
\references{
A wrapper around the brilliant R functions \code{pillar::glimpse()}, \code{dplyr::slice_sample()}, and \code{skimr::skim()}.
}
